Multi-Server Objects Saving Architecture
Because we have all three of our servers writing to one central
repository of database entries we also have objects saving across
servers. This is not ideal. Player saves makes sense to span multiple
servers while objects does not.

With this update I have made it so that when we package our mission it
saves one mission for every server. It also copies a define for the
database entry that corresponds to the server number. This enables us to
keep objects saves segregated across n servers.